[update] 掲示板表示機能をForumControllerに移行
掲示板表示に関する責務を明確にするため、PostControllerから掲示板の一覧表示メソッドを削除し、ForumControllerに移行しました。これにより、コントローラーの責務が分離され、コードの可読性と保守性が向上しました。
